refactoring(utils): reorganization utils
